Split out SerialStringTable from slang-serialize-ir
