+++ source/slang/slang-serialize-debug.cpp
@@ -1,0 +1,13 @@
+// slang-serialize-debug.cpp
+#include "slang-serialize-debug.h"
+
+#include "../core/slang-text-io.h"
+#include "../core/slang-byte-encode-util.h"
+
+#include "slang-ir-insts.h"
+
+#include "../core/slang-math.h"
+
+namespace Slang {
+
+} // namespace Slang
