+++ source/slang/slang-serialize-debug.h
@@ -1,0 +1,16 @@
+// slang-serialize-debug.h
+#ifndef SLANG_SERIALIZE_DEBUG_H
+#define SLANG_SERIALIZE_DEBUG_H
+
+#include "../core/slang-riff.h"
+#include "../core/slang-string-slice-pool.h"
+#include "../core/slang-array-view.h"
+
+#include "slang-name.h"
+#include "slang-source-loc.h"
+
+namespace Slang {
+
+} // namespace Slang
+
+#endif
